requirements.txt
Flask~=2.0.1
Pandas~=1.3.2
Numpy~=1.21.2
PyYAML~=5.4.1
SQLAlchemy~=1.4.23
psycopg2~=2.9.2